+++ models/totaltipe_model.go
@@ -5,6 +5,6 @@ package models
 type Totaltipe struct {
 	Tipe      string  `json:"tipe,omitempty" validate:"required"`
 	Total     float32 `json:"total,omitempty" validate:"required"`
-	Totalpagu float32 `json:"totalpagu,omitempty" validate:"required"`
-	Pdn       float32 `json:"pdn,omitempty" validate:"required"`
+	Totalpagu float64 `json:"totalpagu,omitempty" validate:"required"`
+	Pdn       float64 `json:"pdn,omitempty" validate:"required"`
 }
